tegra: Add function to release resources
Add a companion function to nvhost_client_device_get_resources() called
nvhost_client_device_put_resources() that does the opposite thing.  This
is useful for any nvhost clients that need to be loaded as modules, since
the driver removal path will be exercised in that case.

Signed-off-by: Andrew Chew <achew@nvidia.com>
Change-Id: I72ffc1e8d3eb7bc0d86896e80f121d2a432abbad
Reviewed-on: http://git-master/r/145343
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Allen Martin <amartin@nvidia.com>