ARM: tegra: cpuquiet: make no_lp = -1 always force cluster switch
When setting the no_lp sysfs node to -1 the system would attempt to
switch to the LP cluster. This can fail if the frequency or the number
of online cores are too high.

This change ensure the frequency and number of online cores are first
brought down, such that a cluster switch on no_lp = -1 always
succeeds.

Bug 1247636
Bug 1253920

Change-Id: If6af671e8defdfc8172e8d9356ddf03d54bae98d
Signed-off-by: Peter Boonstoppel <pboonstoppel@nvidia.com>
Reviewed-on: http://git-master/r/217033
Reviewed-by: Diwakar Tundlam <dtundlam@nvidia.com>
Tested-by: Diwakar Tundlam <dtundlam@nvidia.com>
Reviewed-by: Automatic_Commit_Validation_User